Add files via upload
The Hall Booking System project combines essential web development technologies to provide an efficient and user-friendly solution for managing hall bookings. By leveraging HTML, CSS, and JavaScript for the frontend, PHP for the backend, and MySQL for the database, the system ensures a seamless and robust booking experience for users and administrators alike.